Addition of i2c interface (basic version)

+++ src/settings.c
@@ -1,10 +1,10 @@
 #include "settings.h"
 
-uint32_t wheels_gap;
-uint32_t max_linear_acceleration;
-uint32_t max_angular_acceleration;
-uint32_t cruise_linear_speed;
-uint32_t cruise_angular_speed;
-uint32_t ticks_per_cm;
+uint8_t wheels_gap;
+uint16_t max_linear_acceleration;
+uint16_t max_angular_acceleration;
+uint16_t cruise_linear_speed;
+uint16_t cruise_angular_speed;
+uint8_t ticks_per_cm;
 
 uint16_t angular_trust_threshold;
